Show how without the :invalid callback things work
This is because the form, being an OpenStruct, it doesn't implement the
`#invalid?` method `InviteUser` relies on. This makes skip the early
return and normally go on with the invitation, however, the DB
complaints about the violation of the null constraint (the email is
empty) and raises. This is then caught by the `UserProcessor`.